Rework stats interface. Fix WASM/MIME issue. Styling fixes.
