+++ index.css
@@ -11,6 +11,10 @@ html, body {
     height: 100%;
 }
 
+body {
+    overflow: auto;
+}
+
 h1, h2, h3, h4 {
     text-align: center;
 }
@@ -18,7 +22,7 @@ h1, h2, h3, h4 {
 .page-container {
     margin: auto;
     width: 100%;
-    height: 100%;
+    min-height: 100%;
     display: flex;
     gap: 1em;
     flex-direction: column;
@@ -67,6 +71,29 @@ canvas {
     font-size: smaller;
     bottom: 0.2em;
     left: 0.5em;
+}
+
+.stats-table .name {
+    font-size: smaller;
+    text-align: right;
+    padding-right: 0.5em;
+}
+
+.stats-table .value {
+    font-size: smaller;
+    text-align: left;
+}
+
+.stats-container.hidden .stats-table {
+    display: none;
+}
+
+input[type=button].stats-close {
+    font-size: large;
+    text-align: center;
+    width: 100%;
+    min-height: unset;
+    min-width: 7em;
 }
 
 @media (min-width: 600px) {
@@ -120,6 +147,7 @@ td.output {
 button, input[type=button] {
     width: 7em;
     min-height: 3em;
+    cursor: pointer;
 }
 
 input[type=range] {
