+++ utils.js
@@ -1,29 +1,16 @@
-module.exports = function() {
+
+export const toNames = (name) => {
+    const capitalize = (s) => s.substring(0, 1).toUpperCase() + s.substring(1);
+
+    // Convert a string to HTML kebab-cases and javascript camelCases
+    const [first, ...rest] = Array.from(name.matchAll(/([A-Z]*[a-z0-9]+)|([A-Z]+)/g), x => x[0]);
+    const cssWords = [first, ...rest].map((s) => s.toLowerCase());
+    const variableWords = [first.toLowerCase()].concat(rest.map(capitalize));
+    const classWords = [first, ...rest].map(capitalize);
+
     return {
-        getFieldVector: function(x, y) {
-            // Get the field vector and strength at an abritrary position
-            // in the grid
-            vec = {x:0, y:0, strength:0}
-            for (var charge of charges) {
-                dx = (x - charge.x)
-                dx2 = dx * dx;
-                dy = (y - charge.y)
-                dy2 = dy * dy
-                r2 = dx2 + dy2
-                // Check to ensure r2 is > 0
-                if (r2 < 1) r2 = 1
-                
-                // Calculate the field strength due to this charge
-                strength = k * charge.q / (r2)
-        
-                // Add it to the field vectors
-                vec.x += strength * (dx2/r2) * Math.sign(dx)
-                vec.y += strength * (dy2/r2) * Math.sign(dy)
-                vec.strength += strength
-                // console.log(vec)
-            }
-        
-            return vec
-        }
+        'cssName': cssWords.join('-'),
+        'variableName': variableWords.join(''),
+        'className': classWords.join(''),
     }
 }
